94 Glan Road
"In this sinister way I came into my inheritance. You will ask me why I did not dispose of it? I
answer, because I was well convinced that our troubles were in some way dependent upon an incident
in my uncle's life, and that the danger would be as pressing in one house as in another.

"It was in January, '85, that my poor father met his end, and two years and eight months have
elapsed since then. During that time I have lived happily at Horsham, and I had begun to hope that
this curse had passed away from the family, and that it had ended with the last generation. I had
begun to take comfort too soon, however; yesterday morning the blow fell in the very shape in which
it had come upon my father."