84 Wilson Gardens
"'My dearest Uncle: "'I feel that I have brought trouble upon you, and that if I had acted
differently this terrible misfortune might never have occurred. I cannot, with this thought in my
mind, ever again be happy under your roof, and I feel that I must leave you forever. Do not worry
about my future, for that is provided for; and, above all, do not search for me, for it will be
fruitless labour and an ill-service to me. In life or in death, I am ever "'Your loving "'Mary.'

"What could she mean by that note, Mr. Holmes? Do you think it points to suicide?"

"No, no, nothing of the kind. It is perhaps the best possible solution. I trust, Mr. Holder, that
you are nearing the end of your troubles."