66 Tamworth Drive
"It was in the early '60's at the diggings. I was a young chap then, hot-blooded and reckless, ready
to turn my hand at anything; I got among bad companions, took to drink, had no luck with my claim,
took to the bush, and in a word became what you would call over here a highway robber. There were
six of us, and we had a wild, free life of it, sticking up a station from time to time, or stopping
the wagons on the road to the diggings. Black Jack of Ballarat was the name I went under, and our
party is still remembered in the colony as the Ballarat Gang.

"One day a gold convoy came down from Ballarat to Melbourne, and we lay in wait for it and attacked
it. There were six troopers and six of us, so it was a close thing, but we emptied four of their
saddles at the first volley. Three of our boys were killed, however, before we got the swag. I put
my pistol to the head of the wagon-driver, who was this very man McCarthy. I wish to the Lord that I
had shot him then, but I spared him, though I saw his wicked little eyes fixed on my face, as though
to remember every feature. We got away with the gold, became wealthy men, and made our way over to
England without being suspected. There I parted from my old pals and determined to settle down to a
quiet and respectable life. I bought this estate, which chanced to be in the market, and I set
myself to do a little good with my money, to make up for the way in which I had earned it. I
married, too, and though my wife died young she left me my dear little Alice. Even when she was just
a baby her wee hand seemed to lead me down the right path as nothing else had ever done. In a word,
I turned over a new leaf and did my best to make up for the past. All was going well when McCarthy
laid his grip upon me.

"I had gone up to town about an investment, and I met him in Regent Street with hardly a coat to his
back or a boot to his foot.

"'Here we are, Jack,' says he, touching me on the arm; 'we'll be as good as a family to you. There's
two of us, me and my son, and you can have the keeping of us. If you don't--it's a fine, law-abiding
country is England, and there's always a policeman within hail.'

"Well, down they came to the west country, there was no shaking them off, and there they have lived
rent free on my best land ever since. There was no rest for me, no peace, no forgetfulness; turn
where I would, there was his cunning, grinning face at my elbow. It grew worse as Alice grew up, for
he soon saw I was more afraid of her knowing my past than of the police. Whatever he wanted he must
have, and whatever it was I gave him without question, land, money, houses, until at last he asked a
thing which I could not give. He asked for Alice.