80 Corbet Close
"I was aware of it," said Holmes dryly.

"The circumstances are of great delicacy, and every precaution has to be taken to quench what might
grow to be an immense scandal and seriously compromise one of the reigning families of Europe. To
speak plainly, the matter implicates the great House of Ormstein, hereditary kings of Bohemia."

"I was also aware of that," murmured Holmes, settling himself down in his armchair and closing his
eyes.

Our visitor glanced with some apparent surprise at the languid, lounging figure of the man who had
been no doubt depicted to him as the most incisive reasoner and most energetic agent in Europe.
Holmes slowly reopened his eyes and looked impatiently at his gigantic client.

"If your Majesty would condescend to state your case," he remarked, "I should be better able to
advise you."

The man sprang from his chair and paced up and down the room in uncontrollable agitation. Then, with
a gesture of desperation, he tore the mask from his face and hurled it upon the ground. "You are
right," he cried; "I am the King. Why should I attempt to conceal it?"