46 Holmefield Avenue
For answer Holmes clapped the hat upon his head. It came right over the forehead and settled upon
the bridge of his nose. "It is a question of cubic capacity," said he; "a man with so large a brain
must have something in it."

"The decline of his fortunes, then?"