117 Tamworth Drive
"'Photography is one of my hobbies,' said he. 'I have made my dark room up there. But, dear me! what
an observant young lady we have come upon. Who would have believed it? Who would have ever believed
it?' He spoke in a jesting tone, but there was no jest in his eyes as he looked at me. I read
suspicion there and annoyance, but no jest.

"Well, Mr. Holmes, from the moment that I understood that there was something about that suite of
rooms which I was not to know, I was all on fire to go over them. It was not mere curiosity, though
I have my share of that. It was more a feeling of duty--a feeling that some good might come from my
penetrating to this place. They talk of woman's instinct; perhaps it was woman's instinct which gave
me that feeling. At any rate, it was there, and I was keenly on the lookout for any chance to pass
the forbidden door.

"It was only yesterday that the chance came. I may tell you that, besides Mr. Rucastle, both Toller
and his wife find something to do in these deserted rooms, and I once saw him carrying a large black
linen bag with him through the door. Recently he has been drinking hard, and yesterday evening he
was very drunk; and when I came upstairs there was the key in the door. I have no doubt at all that
he had left it there. Mr. and Mrs. Rucastle were both downstairs, and the child was with them, so
that I had an admirable opportunity. I turned the key gently in the lock, opened the door, and
slipped through.

"There was a little passage in front of me, unpapered and uncarpeted, which turned at a right angle
at the farther end. Round this corner were three doors in a line, the first and third of which were
open. They each led into an empty room, dusty and cheerless, with two windows in the one and one in
the other, so thick with dirt that the evening light glimmered dimly through them. The centre door
was closed, and across the outside of it had been fastened one of the broad bars of an iron bed,
padlocked at one end to a ring in the wall, and fastened at the other with stout cord. The door
itself was locked as well, and the key was not there. This barricaded door corresponded clearly with
the shuttered window outside, and yet I could see by the glimmer from beneath it that the room was
not in darkness. Evidently there was a skylight which let in light from above. As I stood in the
passage gazing at the sinister door and wondering what secret it might veil, I suddenly heard the
sound of steps within the room and saw a shadow pass backward and forward against the little slit of
dim light which shone out from under the door. A mad, unreasoning terror rose up in me at the sight,
Mr. Holmes. My overstrung nerves failed me suddenly, and I turned and ran--ran as though some
dreadful hand were behind me clutching at the skirt of my dress. I rushed down the passage, through
the door, and straight into the arms of Mr. Rucastle, who was waiting outside.