158 Sunningdale Drive
Sherlock Holmes laughed. "Here is the foresight," said he putting his finger upon the little disc
and loop of the hat-securer.  "They are never sold upon hats. If this man ordered one, it is a sign
of a certain amount of foresight, since he went out of his way to take this precaution against the
wind. But since we see that he has broken the elastic and has not troubled to replace it, it is
obvious that he has less foresight now than formerly, which is a distinct proof of a weakening
nature. On the other hand, he has endeavoured to conceal some of these stains upon the felt by
daubing them with ink, which is a sign that he has not entirely lost his self-respect."

"Your reasoning is certainly plausible."