99 Holmefield Avenue
"Yes, and a most remarkable bird it proved. I don't wonder that you should take an interest in it.
It laid an egg after it was dead--the bonniest, brightest little blue egg that ever was seen. I have
it here in my museum."

Our visitor staggered to his feet and clutched the mantelpiece with his right hand. Holmes unlocked
his strong-box and held up the blue carbuncle, which shone out like a star, with a cold, brilliant,
many-pointed radiance. Ryder stood glaring with a drawn face, uncertain whether to claim or to
disown it.

"The game's up, Ryder," said Holmes quietly. "Hold up, man, or you'll be into the fire! Give him an
arm back into his chair, Watson. He's not got blood enough to go in for felony with impunity. Give
him a dash of brandy. So! Now he looks a little more human. What a shrimp it is, to be sure!"