10 Larch Walk
"I am all attention, madam."

"My name is Helen Stoner, and I am living with my stepfather, who is the last survivor of one of the
oldest Saxon families in England, the Roylotts of Stoke Moran, on the western border of Surrey."

Holmes nodded his head. "The name is familiar to me," said he.

"The family was at one time among the richest in England, and the estates extended over the borders
into Berkshire in the north, and Hampshire in the west. In the last century, however, four
successive heirs were of a dissolute and wasteful disposition, and the family ruin was eventually
completed by a gambler in the days of the Regency. Nothing was left save a few acres of ground, and
the two-hundred-year-old house, which is itself crushed under a heavy mortgage. The last squire
dragged out his existence there, living the horrible life of an aristocratic pauper; but his only
son, my stepfather, seeing that he must adapt himself to the new conditions, obtained an advance
from a relative, which enabled him to take a medical degree and went out to Calcutta, where, by his
professional skill and his force of character, he established a large practice. In a fit of anger,
however, caused by some robberies which had been perpetrated in the house, he beat his native butler
to death and narrowly escaped a capital sentence. As it was, he suffered a long term of imprisonment
and afterwards returned to England a morose and disappointed man.

"When Dr. Roylott was in India he married my mother, Mrs. Stoner, the young widow of Major-General
Stoner, of the Bengal Artillery. My sister Julia and I were twins, and we were only two years old at
the time of my mother's re-marriage. She had a considerable sum of money--not less than £1000 a year
--and this she bequeathed to Dr. Roylott entirely while we resided with him, with a provision that a
certain annual sum should be allowed to each of us in the event of our marriage. Shortly after our
return to England my mother died--she was killed eight years ago in a railway accident near Crewe.
Dr. Roylott then abandoned his attempts to establish himself in practice in London and took us to
live with him in the old ancestral house at Stoke Moran. The money which my mother had left was
enough for all our wants, and there seemed to be no obstacle to our happiness.