130 Glan Road
"My heart had turned to lead. 'It is K. K. K.,' said I.

"He looked inside the envelope. 'So it is,' he cried. 'Here are the very letters. But what is this
written above them?'

"'Put the papers on the sundial,' I read, peeping over his shoulder.

"'What papers? What sundial?' he asked.

"'The sundial in the garden. There is no other,' said I; 'but the papers must be those that are
destroyed.'