35 Glan Road
"The doctor?"

"Yes, have you not heard? Poor father has never been strong for years back, but this has broken him
down completely. He has taken to his bed, and Dr. Willows says that he is a wreck and that his
nervous system is shattered. Mr. McCarthy was the only man alive who had known dad in the old days
in Victoria."

"Ha! In Victoria! That is important."

"Yes, at the mines."