115 Beaconside
"Yes, to-day." She stood smiling, holding up a little slip of paper in the air.

"May I see it?"

"Certainly."

He snatched it from her in his eagerness, and smoothing it out upon the table he drew over the lamp
and examined it intently. I had left my chair and was gazing at it over his shoulder. The envelope
was a very coarse one and was stamped with the Gravesend postmark and with the date of that very
day, or rather of the day before, for it was considerably after midnight.