44 Harvey Street
"I will try. What is it?"

"We shall be at the Copper Beeches by seven o'clock, my friend and I. The Rucastles will be gone by
that time, and Toller will, we hope, be incapable. There only remains Mrs. Toller, who might give
the alarm. If you could send her into the cellar on some errand, and then turn the key upon her, you
would facilitate matters immensely."

"I will do it."

"Excellent! We shall then look thoroughly into the affair. Of course there is only one feasible
explanation. You have been brought there to personate someone, and the real person is imprisoned in
this chamber. That is obvious. As to who this prisoner is, I have no doubt that it is the daughter,
Miss Alice Rucastle, if I remember right, who was said to have gone to America. You were chosen,
doubtless, as resembling her in height, figure, and the colour of your hair. Hers had been cut off,
very possibly in some illness through which she has passed, and so, of course, yours had to be
sacrificed also. By a curious chance you came upon her tresses. The man in the road was undoubtedly
some friend of hers--possibly her fiancé--and no doubt, as you wore the girl's dress and were so
like her, he was convinced from your laughter, whenever he saw you, and afterwards from your
gesture, that Miss Rucastle was perfectly happy, and that she no longer desired his attentions. The
dog is let loose at night to prevent him from endeavouring to communicate with her. So much is
fairly clear. The most serious point in the case is the disposition of the child."

"What on earth has that to do with it?" I ejaculated.