6 Badgers Dene
"Between nine and ten last night Police-Constable Cook, of the H Division, on duty near Waterloo
Bridge, heard a cry for help and a splash in the water. The night, however, was extremely dark and
stormy, so that, in spite of the help of several passers-by, it was quite impossible to effect a
rescue. The alarm, however, was given, and, by the aid of the water-police, the body was eventually
recovered. It proved to be that of a young gentleman whose name, as it appears from an envelope
which was found in his pocket, was John Openshaw, and whose residence is near Horsham. It is
conjectured that he may have been hurrying down to catch the last train from Waterloo Station, and
that in his haste and the extreme darkness he missed his path and walked over the edge of one of the
small landing-places for river steamboats. The body exhibited no traces of violence, and there can
be no doubt that the deceased had been the victim of an unfortunate accident, which should have the
effect of calling the attention of the authorities to the condition of the riverside landing-
stages."

We sat in silence for some minutes, Holmes more depressed and shaken than I had ever seen him.

"That hurts my pride, Watson," he said at last. "It is a petty feeling, no doubt, but it hurts my
pride. It becomes a personal matter with me now, and, if God sends me health, I shall set my hand
upon this gang. That he should come to me for help, and that I should send him away to his death--!"
He sprang from his chair and paced about the room in uncontrollable agitation, with a flush upon his
sallow cheeks and a nervous clasping and unclasping of his long thin hands.

"They must be cunning devils," he exclaimed at last. "How could they have decoyed him down there?
The Embankment is not on the direct line to the station. The bridge, no doubt, was too crowded, even
on such a night, for their purpose. Well, Watson, we shall see who will win in the long run. I am
going out now!"

"To the police?"

"No; I shall be my own police. When I have spun the web they may take the flies, but not before."