101 Wilson Gardens
"Oh, it is a matter of no importance; but I don't know why you should be so warm over such a
trifle."

"Warm! You'd be as warm, maybe, if you were as pestered as I am. When I pay good money for a good
article there should be an end of the business; but it's 'Where are the geese?' and 'Who did you
sell the geese to?' and 'What will you take for the geese?' One would think they were the only geese
in the world, to hear the fuss that is made over them."