42 Bowler Street
"But it might have been a cry of surprise. Astonishment at the unexpected sight of you might cause
him to throw up his hands?"

"It is possible."

"And you thought he was pulled back?"

"He disappeared so suddenly."

"He might have leaped back. You did not see anyone else in the room?"

"No, but this horrible man confessed to having been there, and the Lascar was at the foot of the
stairs."