161 Sunningdale Drive
"My God! It's Watson," said he. He was in a pitiable state of reaction, with every nerve in a
twitter. "I say, Watson, what o'clock is it?"

"Nearly eleven."

"Of what day?"

"Of Friday, June 19th."

"Good heavens! I thought it was Wednesday. It is Wednesday. What d'you want to frighten a chap for?"
He sank his face onto his arms and began to sob in a high treble key.

"I tell you that it is Friday, man. Your wife has been waiting this two days for you. You should be
ashamed of yourself!"