64 Cliff Place
"He is still with you, I presume?"

"Yes, sir. He and a girl of fourteen, who does a bit of simple cooking and keeps the place clean--
that's all I have in the house, for I am a widower and never had any family. We live very quietly,
sir, the three of us; and we keep a roof over our heads and pay our debts, if we do nothing more.

"The first thing that put us out was that advertisement. Spaulding, he came down into the office
just this day eight weeks, with this very paper in his hand, and he says:

"'I wish to the Lord, Mr. Wilson, that I was a red-headed man.'

"'Why that?' I asks.