96 Glan Road
The banker's dressing-room was a plainly furnished little chamber, with a grey carpet, a large
bureau, and a long mirror. Holmes went to the bureau first and looked hard at the lock.

"Which key was used to open it?" he asked.

"That which my son himself indicated--that of the cupboard of the lumber-room."