39 Wellswood Gardens
The young man took from his waistcoat a crumpled envelope, and turning to the table he shook out
upon it five little dried orange pips.

"This is the envelope," he continued. "The postmark is London--eastern division. Within are the very
words which were upon my father's last message: 'K. K. K.'; and then 'Put the papers on the
sundial.'"