37 Badgers Dene
"Done about the same time as the bell-rope?" remarked Holmes.

"Yes, there were several little changes carried out about that time."

"They seem to have been of a most interesting character--dummy bell-ropes, and ventilators which do
not ventilate. With your permission, Miss Stoner, we shall now carry our researches into the inner
apartment."

Dr. Grimesby Roylott's chamber was larger than that of his step-daughter, but was as plainly
furnished. A camp-bed, a small wooden shelf full of books, mostly of a technical character, an
armchair beside the bed, a plain wooden chair against the wall, a round table, and a large iron safe
were the principal things which met the eye. Holmes walked slowly round and examined each and all of
them with the keenest interest.