58 Glan Road
"And now I will show you what is really a very interesting study, Mr. Windibank," Holmes continued.
"I think of writing another little monograph some of these days on the typewriter and its relation
to crime. It is a subject to which I have devoted some little attention. I have here four letters
which purport to come from the missing man. They are all typewritten. In each case, not only are the
'e's' slurred and the 'r's' tailless, but you will observe, if you care to use my magnifying lens,
that the fourteen other characteristics to which I have alluded are there as well."

Mr. Windibank sprang out of his chair and picked up his hat. "I cannot waste time over this sort of
fantastic talk, Mr. Holmes," he said. "If you can catch the man, catch him, and let me know when you
have done it."

"Certainly," said Holmes, stepping over and turning the key in the door. "I let you know, then, that
I have caught him!"

"What! where?" shouted Mr. Windibank, turning white to his lips and glancing about him like a rat in
a trap.

"Oh, it won't do--really it won't," said Holmes suavely. "There is no possible getting out of it,
Mr. Windibank. It is quite too transparent, and it was a very bad compliment when you said that it
was impossible for me to solve so simple a question. That's right! Sit down and let us talk it
over."