76 Badgers Dene
"Well, you can easily think that that made me prick up my ears, for the business has not been over-
good for some years, and an extra couple of hundred would have been very handy.

"'Tell me all about it,' said I.

"'Well,' said he, showing me the advertisement, 'you can see for yourself that the League has a
vacancy, and there is the address where you should apply for particulars. As far as I can make out,
the League was founded by an American millionaire, Ezekiah Hopkins, who was very peculiar in his
ways. He was himself red-headed, and he had a great sympathy for all red-headed men; so when he died
it was found that he had left his enormous fortune in the hands of trustees, with instructions to
apply the interest to the providing of easy berths to men whose hair is of that colour. From all I
hear it is splendid pay and very little to do.'

"'But,' said I, 'there would be millions of red-headed men who would apply.'

"'Not so many as you might think,' he answered. 'You see it is really confined to Londoners, and to
grown men. This American had started from London when he was young, and he wanted to do the old town
a good turn. Then, again, I have heard it is no use your applying if your hair is light red, or dark
red, or anything but real bright, blazing, fiery red. Now, if you cared to apply, Mr. Wilson, you
would just walk in; but perhaps it would hardly be worth your while to put yourself out of the way
for the sake of a few hundred pounds.'