154 Glan Road
"I should be immensely obliged to you."

"We'll call a cab and go together. We shall just be in time to have a little breakfast with him. Do
you feel equal to it?"

"Yes; I shall not feel easy until I have told my story."

"Then my servant will call a cab, and I shall be with you in an instant." I rushed upstairs,
explained the matter shortly to my wife, and in five minutes was inside a hansom, driving with my
new acquaintance to Baker Street.

Sherlock Holmes was, as I expected, lounging about his sitting-room in his dressing-gown, reading
the agony column of The Times and smoking his before-breakfast pipe, which was composed of all the
plugs and dottles left from his smokes of the day before, all carefully dried and collected on the
corner of the mantelpiece. He received us in his quietly genial fashion, ordered fresh rashers and
eggs, and joined us in a hearty meal. When it was concluded he settled our new acquaintance upon the
sofa, placed a pillow beneath his head, and laid a glass of brandy and water within his reach.